Queue Family Indicies Done
Rearanged queuefamily indices to match better with what's should be
done, for the moment looks exatcly like the vulkantutorial structure but
I might end up doing a custom structure with enum to get each values
instead of questing a field in a structure.
TODO : Better function to select the most appropriate Physical Device.