Starting Documentation and SwapchainManagerProgress
Some class and stucture have recieved Doxygen comments plus add a
Doxyfile to generate documentation.
TODO : add Doxygen to the cmake generation process.
TODO : Finish the SwapchainManager and add it to the ressourceManager as
a vector + add in the ressource manager a vector of these elements.
Might want to handle device allocation when we don't want presentation
or graphics.